refactor: update coding style as per codacy
