+++ ui_tests/conftest.py
@@ -6,7 +6,7 @@ import pytest
 from playwright.sync_api import Error as PlaywrightError
 from playwright.sync_api import Page, Playwright, expect
 
-from ui_tests.pages.form_page import form_page  # noqa: F401
+from ui_tests.pages.form_page import FormPage
 
 BENCH_PATH = Path(__file__).parents[3]
 # Rewritten on every run, and left on disk so `playwright codegen --load-storage`
@@ -154,3 +154,14 @@ def doc(site):
             print(f"could not clean up {doctype} {name}: {error}")
 
     frappe.db.commit()  # nosemgrep
+
+
+@pytest.fixture(autouse=True)
+def form_page(request, authenticated_desk: Page, doc):
+    def open_form(doctype: str, name: str | None = None) -> FormPage:
+        return FormPage(authenticated_desk, doctype, name, track=doc).navigate()
+
+    if request.instance is not None:
+        request.instance.form_page = open_form
+
+    return open_form
